# spring-server/src/main/resources/api-spec.yaml
openapi: 3.0.1
info:
  title: Articles app API 
  description: "API de mon grand projet qui n'a pas encore de nom. App sur laquelle on peut poster des articles et intéragir avec ceux-ci."
  version: 1.0.0
servers:
- url: https://localhost:8080/api
...
paths:
  /articles:
    get:
      tags:
      - articles
      summary: Finds all articles by date
      operationId: findArticlesByDate
      responses:
        200: 
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        404:
          description: Articles not found
          content: {}
    post:
      tags:
      - articles
      summary: Creates a new article
      operationId: createArticle
      requestBody:
        description: Article that needs to be added
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        405:
          description: Invalid input
          content: {}
  /articles/byCategories:
    get:
      tags:
      - articles
      summary: Finds articles by categories
      operationId: findArticlesByCategories
      parameters:
      - name: category
        in: query
        description: Categories to filter by
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            $ref: '#/components/schemas/Category'
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Article'
        400:
          description: Invalid category 
          content: {}
  /articles/{articleId}:
    get:
      tags:
      - articles
      summary: Find article by ID
      description: Returns a single article by ID
      operationId: FindArticleById
      parameters:
      - name: articleId
        in: path
        description: ID of the article to return
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Article'
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Article not found
          content: {}
    patch:
      tags:
      - articles
      summary: Updates an article by ID
      operationId: updateArticleById
      parameters:
      - name: articleId
        in: path
        description: ID of article that needs to be updated
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: updated article
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Article'
        required: true
      responses:
        200: 
          description: updated successfully
        405:
          description: Invalid input
          content: {}
    delete:
      tags:
      - articles
      summary: Deletes an article by ID
      operationId: deleteArticleById
      parameters:
      - name: articleId
        in: path
        description: Article id to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200: 
          description: deleted successfully
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Article not found
          content: {}
  /articles/{articleId}/comments:
    get:
      tags:
      - comments
      summary: returns an array of comments for a particular article
      operationId: findCommentsByArticle
      parameters:
      - name: articleId
        in: path
        description: Id of the article
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        400:
          description: Invalid tag value
          content: {}
    post:
      tags:
      - comments
      summary: add a new comment to an article
      operationId: addCommentToAnArticle
      parameters:
      - name: commentId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: articleId
        in: path
        description: ID of the article
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: comment to add to the article
        content: 
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
          200:
            description: added successfully
          405:
            description: Invalid input
            content: {} 
  /articles/{articleId}/comments/{commentId}:
    get:
      tags:
      - comments
      summary: returns a comment by ID
      operationId: findCommentbyID
      parameters:
      - name: commentId
        in: path
        description: ID of the comment
        required: true
        schema:
          type: integer
          format: int64
      - name: articleId
        in: path
        description: ID of the article
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Comment'
        400:
          description: Invalid tag value
          content: {}
    patch:
      tags:
      - comments
      summary: Updates a comment by ID
      operationId: updateCommentbyId
      parameters:
      - name: commentId
        in: path
        required: true
        description: ID of the comment that needs to be updated
        schema:
          type: integer
          format: int64
      - name: articleId
        in: path
        description: ID of the article
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        description: updated comment
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Comment'
      responses:
        200:
          description: updated successfully
        405:
          description: Invalid input
          content: {}
    delete:
      tags:
      - comments
      summary: Deletes a comment by ID
      operationId: deleteComment
      parameters:
      - name: commentId
        in: path
        description: ID of the comment to delete
        required: true
        schema:
          type: integer
          format: int64
      - name: articleId
        in: path
        description: Article ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        200:
          description: deleted successfully
        400:
          description: Invalid ID supplied
          content: {}
        404:
          description: Comment not found
          content: {}
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        password:
          type: string
    Tag:
      required:
        - id 
        - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    Article:
      required:
      - id
      - title
      - tags
      - content
      - author
      - createdAt
      - lastUpdatedAt
      - views
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: my super article
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        content:
          type: string
          example: "Le Lorem Ipsum est simplement du faux texte employé dans la composition et la mise en page avant impression. Le Lorem Ipsum est le faux texte standard de l'imprimerie depuis les années 1500..."
        photoUrls:
          type: array
          items:
            type: string
        author: 
          $ref: '#/components/schemas/User'
        createdAt: 
          type: string 
          example: '12-02-2013H12-05'
        lastUpdateAt:
          type: string
          example: '12-02-2013H12-05'
        views:
          type: integer
          example: 456
    Comment:
      required: 
        - content
        - article
        - title
        - id 
        - author
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
          example: article nul
        content:
          type: string
          example: tu pourrais faire mieux
        author:
          $ref: '#/components/schemas/User'
        article: 
          $ref: '#/components/schemas/Article'
    Category:
      required:
        - name
        - children
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        children:
          type: array
          items:
            $ref: '#/components/schemas/Tag'